# Fixed #15675 - Improve report expiry performance