@lzap 2016-07-21
I removed the query plan output from the commit message telling to visit this PR to see them. Can you please reformat them into nicer format if you happen to have them? It's all hard-wrapped. If not, no big deal.